Scope authenticated custom routes by tenant
